+++ src/TranslationsConfiguration.tsx
@@ -2,16 +2,13 @@ import React, { useEffect, useState } from "react";
 import { ReactNode } from "react";
 import RestClient from "./RestClient";
 import { useRestClientConfiguration } from "./RestClientConfiguration";
+import { RetrieveTranslations } from "./RetrieveTranslations";
 
-interface TranslationMap {
+export interface TranslationMap {
     [key: string]: string;
 }
 
 type TranslationRepository = (key: string) => string;
-
-interface RemoteTranslations {
-    translations: TranslationMap;
-}
 
 interface Translations {
     data: TranslationMap;
@@ -23,23 +20,6 @@ const retrieveConstTranslations = (): TranslationMap => {
         'appflow.customerData.hi': 'Hi',
         'appflow.customerData.t_and_c': 'Accept t&c',
     }
-}
-
-const RetrieveTranslations = (restClient: RestClient): TranslationMap => {
-    const [data, setData] = useState<TranslationMap>({} as TranslationMap);
-
-
-    useEffect(()=> {
-        const response = restClient.get<RemoteTranslations>('/translations/en');
-
-        response.then((translations) =>{
-            setData(translations.translations);
-        }).catch(()=> {
-            console.log('Error fetching translations');
-        })
-    }, [])
-
-    return data;
 }
 
 const createTranslationRepository = (data: TranslationMap): TranslationRepository => {
